requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-timezone-field==7.1
django-treebeard==4.7.1
djangorestframework==3.16.0
kombu==5.5.3
numpy==2.2.5
paho-mqtt==2.1.0
prompt_toolkit==3.0.51
psycopg==3.2.7
psycopg-binary==3.2.7
pyhumps==3.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==6.0.0
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
